requirements.txt
librosa==0.11.0
torch==2.7.0
torchaudio==2.9.0
muq==0.1.0
hydra-core==1.3.2